backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
pandas==2.2.2
prophet==1.1.5
cmdstanpy==1.2.4
pywebview==4.4.1
yfinance==0.2.40

